ISM6236 WCF Services Example
This was the final assignment written for ISM6236 in the C# language utilizing the dblib.dll code (ADO.NET) from a previous assignment.  The program demonstrates the SOAP protocol with the creations of the WSDL document from a WCF Services project.  The client (a console application) creates a stub to call methods from the service to access the methods that link to the database.