IPS Automation GUI Tool
This is a demonstration of an Investment Policy Statement Automation GUI Tool that will generate multiple IPSs to the correct clients via a list of account numbers provided by the user.  Alternately, the user can search up an account with the Quick Search feature using several filter options or create a new IPS with the Single IPS tab.  The application has a settings tab where certain file paths and IPS Template references can be applied and saved to a .txt file for future use (should this save file be deleted or end up corrupted, a new save file should autogenerate with the default paths and values at the application's startup).