Stock Screener
This is a work in progress side project written in Python.  The end result should pull stocks filtered based on a variety of attributes that meet certain thresholds and come from sectors that have the lowest relative Schiller PE Ratio to historic average (basically identifying undervalued sectors) that have a net positive money inflow into the sector.  The program currently does this as well as identifies stocks level lines (support and resistance levels).  If a stock's current price is approaching one of these levels, a variety of strategies can be implemented.  The end goal is to have the script notify the user via text or email when the stock poses a buy or sell opportunity.