Quiz Android App
The following is written in Kotlin.  The files contain the .kt Kotlin code files, but does not contain all the files associated with the app since there are too many files to upload at once.  Please reference the QuizApp.zip to see all the files (The files not in the zip are just from the src folder).